Test Fixes
1. Move SSLEventsHelper to test/helper.rb, fixes `warning: method redefined`
2. Collapse 'Debugging Info' in Actions log
3. Fix 'shadowed cariable' warning
