+++ test/test_puma_localhost_authority.rb
@@ -7,18 +7,7 @@ require "localhost/authority"
 
 if ::Puma::HAS_SSL && !Puma::IS_JRUBY
   require "puma/minissl"
-  require "puma/events"
   require "net/http"
-
-  class SSLEventsHelper < ::Puma::Events
-    attr_accessor :addr, :cert, :error
-
-    def ssl_error(error, ssl_socket)
-      self.error = error
-      self.addr = ssl_socket.peeraddr.last rescue "<unknown>"
-      self.cert = ssl_socket.peercert
-    end
-  end
 
   # net/http (loaded in helper) does not necessarily load OpenSSL
   require "openssl" unless Object.const_defined? :OpenSSL
